fix getmempooltxs

+++ service/api/app/internal/repo/mempool/api.go
@@ -8,7 +8,7 @@ import (
 )
 
 type Mempool interface {
-	GetMempoolTxs(offset int64, limit int64) (mempoolTx []*mempoolModel.MempoolTx, err error)
+	GetMempoolTxs(offset int, limit int) (mempoolTx []*mempoolModel.MempoolTx, err error)
 	GetMempoolTxsTotalCount() (count int64, err error)
 	GetMempoolTxsTotalCountByAccountIndex(accountIndex int64) (count int64, err error)
 	GetMempoolTxByTxHash(hash string) (mempoolTxs *mempoolModel.MempoolTx, err error)
